Remove slushy infrastructure from a project.

Deletes the files and folders needed for the slushy environment before restarting.

Usage

slushy_remove(project = NULL, restart = TRUE, keep = "slushy_config.yml")

Arguments

project

The project directory. If NULL, defaults to the nearest parent directory that contains an `.Rproj` file relative to the current working directory.

restart

Restart session after slushy is removed? Defaults to TRUE.

keep

A character vector specifying the names of files to retain during the cleanup process. By default, "slushy_config.yml" is kept.

Value

This function is called for its side effects.

Examples

if (FALSE) { # \dontrun{
 slushy_remove()
} # }