Check for differences between local environment and the project lock file.

Check for differences between local environment and the project lock file.

Usage

slushy_status(project = NULL, pkg_deps_ok = get_config()$pkg_deps_ok)

Arguments

project

The project directory. If NULL, defaults to the nearest parent directory that contains an `.Rproj` file relative to the current working directory.

pkg_deps_ok

Include package dependencies as part of the agreed upon set? If FALSE, any use of these dependencies found in the project will be flagged in a message.

Value

Sync status (TRUE = synced, FALSE = not synced), invisibly. This function is called for its side effects.

Details

The following will be checked for:

- Packages used in files (via library("package") or package::) that are not documented in the lockfile.

- Installed package versions differ from lockfile.

- Packages in lockfile are not installed in local environment.

Examples

if (FALSE) { # \dontrun{
  slushy_status()
} # }